requirements.txt
pandas
requests
pycryptodome
pyinstaller